+++ test/utils_test.exs
@@ -13,4 +13,28 @@ defmodule Maru.UtilsTest do
     assert "aa_bb" == lower_underscore "AaBb"
     assert "ab"    == lower_underscore "Ab"
   end
+
+  test "make_validator" do
+    assert_raise Maru.Exceptions.UndefinedValidator, fn ->
+      make_validator(:a)
+    end
+
+    defmodule Elixir.Maru.Validations.V do
+    end
+
+    assert Elixir.Maru.Validations.V = make_validator(:v)
+  end
+
+  test "make_type" do
+    type = quote do: A
+    assert_raise Maru.Exceptions.UndefinedType, fn ->
+      make_type(type)
+    end
+
+    defmodule Elixir.Maru.Types.B do
+      use Elixir.Maru.Type
+    end
+    type = quote do: B
+    assert Elixir.Maru.Types.B = make_type(type)
+  end
 end